Switch to 64-bit ISD-AS identifiers.
Also:
- Include raw scion packet in wireshark plugin, to allow for easier
  extraction.
- Fix some bugs in hpkt.WriteScnPkt:
  - L4 payload length wasn't set.
  - regBaseHeaderLen was incorrectly calculated, and only accidentally
    correct until now.
- Add ./tools/pktcap to save raw scion packets from the network to a
  file.
- Use real scion packets for go/border/rpkt and go/lib/hpkt unit tests,
  and document how to generate/capture them.
- Remove ancient code from libscion that dates from the pre-capnp era.
- Make hpkt's write test a lot less painful to maintain, by having it
  write and then parse the packet, rather than having to hand edit a
  binary representation.
- Use strconv.ParseUint instead of strconv.Atoi when parsing ISDs in Go,
  for greater control.
- Update various go data structures to use the new addr.ISD/adddr.AS
  types (e.g. TRCs and Certificates)

+++ proto/drkey_mgmt.capnp
@@ -5,7 +5,7 @@ $Go.package("proto");
 $Go.import("github.com/scionproto/scion/go/proto");
 
 struct DRKeyReq {
-    isdas @0 :UInt32;      # Src ISD-AS of the requested DRKey
+    isdas @0 :UInt64;      # Src ISD-AS of the requested DRKey
     timestamp @1 :UInt64;  # Timestamp
     signature @2 :Data;    # Signature of (isdas, prefetch, timestamp)
     certVer @3 :UInt32;    # Version cert used to sign
@@ -17,7 +17,7 @@ struct DRKeyReq {
 }
 
 struct DRKeyRep {
-    isdas @0 :UInt32;      # Src ISD-AS of the DRKey
+    isdas @0 :UInt64;      # Src ISD-AS of the DRKey
     timestamp @1 :UInt64;  # Timestamp
     expTime @2 :UInt64;    # Expiration time of the DRKey
     cipher @3 :Data;       # Encrypted DRKey
