+++ proto/rev_info.capnp
@@ -15,7 +15,7 @@ struct RevInfo {
 	siblings @3 :List(SiblingHash);  # Hash values of siblings, bottom to top
 	prevRoot @4 :Data;  # Root of the hashtree of previous time block (T-1)
 	nextRoot @5 :Data;  # Root of the hashtree of next time block (T+1)
-	isdas @6 :UInt32;  # ISD-AS of the revocation issuer.
+	isdas @6 :UInt64;  # ISD-AS of the revocation issuer.
 	hashType @7 :UInt16;  # The hash function type needed to verify the revocation.
 	treeTTL @8 :UInt32;  # The validity period of the revocation tree.
 }
